+++ src/handler/links.rs
@@ -1,3 +1,4 @@
+use super::filter::filtered_url;
 use super::{log_error, Handler};
 
 use crate::db::DB;
@@ -7,138 +8,8 @@ use crate::structs::Link;
 use chrono::Duration;
 use lazy_static::lazy_static;
 use linkify::{LinkFinder, LinkKind};
-use phf::phf_set;
 use regex::Regex;
 use serenity::{model::channel::Message, prelude::*};
-use url::Url;
-
-// largely sourced from newhouse/url-tracking-stripper on github
-
-static TWITTER_FIELDS: phf::Set<&'static str> = phf_set! {
-    "s",
-    "t"
-};
-
-static GENERIC_FIELDS: phf::Set<&'static str> = phf_set! {
-    // Google's Urchin Tracking Module
-    "utm_source",
-    "utm_medium",
-    "utm_term",
-    "utm_campaign",
-    "utm_content",
-    "utm_name",
-    "utm_cid",
-    "utm_reader",
-    "utm_viz_id",
-    "utm_pubreferrer",
-    "utm_swu",
-    // Mailchimp
-    "mc_cid",
-    "mc_eid",
-    // comScore Digital Analytix?
-    // http://www.about-digitalanalytics.com/comscore-digital-analytix-url-campaign-generator
-    "ns_source",
-    "ns_mchannel",
-    "ns_campaign",
-    "ns_linkname",
-    "ns_fee",
-    // Simple Reach
-    "sr_share",
-    // Facebook Click Identifier
-    // http://thisinterestsme.com/facebook-fbclid-parameter/
-    "fbclid",
-    // Instagram Share Identifier
-    "igshid",
-    "srcid",
-    // Google Click Identifier
-    "gclid",
-    // Some other Google Click thing
-    "ocid",
-    // Unknown
-    "ncid",
-    // Unknown
-    "nr_email_referer",
-    // Generic-ish. Facebook, Product Hunt and others
-    "ref",
-    // Alibaba-family 'super position model' tracker:
-    // https://github.com/newhouse/url-tracking-stripper/issues/38
-    "spm",
-};
-
-/// filter_field returns true if we should filter a field out in a query string,
-/// otherwise returns false.
-///
-/// We filter fields that are largely meant for tracking and as such not meaningfully
-/// useful for comparison purposes. Without filtering out tracking filters otherwise
-/// identical links may not be the same because of different tracking values for
-/// different users.
-///
-/// Requires the host as well as sometimes we do specific filters for specifics hosts
-/// i.e we filter "s" on twitter but nothing else. It should be expected that this
-/// function will grow over time
-#[inline(always)]
-fn filter_field(host: &str, field: &str) -> bool {
-    let host_match = match host {
-        "twitter" | "twitter.com" => TWITTER_FIELDS.contains(field),
-        _ => false,
-    };
-    host_match || GENERIC_FIELDS.contains(field)
-}
-
-fn transform_url(url: Url) -> Result<Url> {
-    let ret = if url.host_str().is_some() {
-        match url.host_str().unwrap() {
-            "youtu.be" => {
-                let path = url.path();
-                if path.len() > 1 {
-                    let expanded_url =
-                        format!("https://www.youtube.com/watch?v={}", &path[1..path.len()]);
-                    Some(Url::parse(&expanded_url)?)
-                } else {
-                    None
-                }
-            }
-            _ => None,
-        }
-    } else {
-        None
-    };
-
-    match ret {
-        Some(value) => Ok(value),
-        None => Ok(url),
-    }
-}
-
-/// filtered_url takes a url_str and returns a Url object with the any irrelevent
-/// fields in the query string removed as per filter_field
-fn filtered_url(url_str: &str) -> Result<Url> {
-    let mut url = transform_url(Url::parse(url_str)?)?;
-    let host = url.host_str().ok_or(rusqlite::Error::QueryReturnedNoRows)?;
-
-    let fields = url
-        .query_pairs()
-        .filter(|(field, _value)| !filter_field(host, field))
-        .map(|(f, v)| (f.into_owned(), v.into_owned()))
-        .collect::<Vec<(String, String)>>();
-
-    let mut query = url.query_pairs_mut();
-    query.clear();
-    for field in fields {
-        query.append_pair(&field.0, &field.1);
-    }
-
-    // need this to ensure no dangling references
-    drop(query);
-
-    // if query is some(empty string) then the result will contain a dangling ?, this removes that
-    if url.query() == Some("") {
-        url.set_query(None);
-    }
-
-    println!("Filtered url: {:?}", url);
-    Ok(url)
-}
 
 fn query_link_matches(url_str: &str, server: u64) -> Result<Vec<Link>> {
     let mut links = Vec::new();
